requirements.txt
loguru==0.7.2
numpy==1.26.4
pillow==10.3.0
PyYAML==6.0.1
scikit-image==0.21.0
Scrapy==2.13.2
selenium==4.33.0
SQLAlchemy==2.0.32
